fix(review): 3 additional findings from code review angle A
- section-bg.php: remove double esc_url() on spray preset URL — already
  escaped by kc_section_bg_build_style(); double encoding corrupts &amp; in CDN URLs
- blocks.php: allow_null=false on spray select (has explicit empty choice)
- galerie.php: cast get_post_thumbnail_id() to int to handle string '0'

+++ template-parts/layouts/galerie.php
@@ -45,7 +45,7 @@ if ( empty( $photos ) ) {
 		]
 	);
 	foreach ( $foto_posts as $foto ) {
-		$img_id = get_post_thumbnail_id( $foto );
+		$img_id = (int) get_post_thumbnail_id( $foto );
 		if ( ! $img_id ) {
 			continue;
 		}
